[ResponseOps][Rules] Move metric rule params schema to package (#205492)
Connected with #195189

- Moved params of duration metric inventory threshold rule type to
`/response-ops/rule_params/metric_inventory_threshold/`
- Moved params of metric threshold rule type to
`/response-ops/rule_params/metric_threshold/`

**I did NOT move the corresponding type to the rule_params package due
to the recursive imports it would create.**

---------

(cherry picked from commit 4873fa18d70142c5c2e6633d51648c45f6f481cb)
