+++ src/platform/packages/shared/response-ops/rule_params/common/utils.ts
@@ -1,0 +1,41 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the "Elastic License
+ * 2.0", the "GNU Affero General Public License v3.0 only", and the "Server Side
+ * Public License v 1"; you may not use this file except in compliance with, at
+ * your election, the "Elastic License 2.0", the "GNU Affero General Public
+ * License v3.0 only", or the "Server Side Public License, v 1".
+ */
+
+import { schema } from '@kbn/config-schema';
+import { isEmpty } from 'lodash';
+
+export const oneOfLiterals = (arrayOfLiterals: Readonly<string[]>) =>
+  schema.string({
+    validate: (value) =>
+      arrayOfLiterals.includes(value) ? undefined : `must be one of ${arrayOfLiterals.join(' | ')}`,
+  });
+
+export const validateIsStringElasticsearchJSONFilter = (value: string) => {
+  if (value === '') {
+    // Allow clearing the filter.
+    return;
+  }
+
+  const errorMessage = 'filterQuery must be a valid Elasticsearch filter expressed in JSON';
+  try {
+    const parsedValue = JSON.parse(value);
+    if (!isEmpty(parsedValue.bool)) {
+      return undefined;
+    }
+    return errorMessage;
+  } catch (e) {
+    return errorMessage;
+  }
+};
+
+export type TimeUnitChar = 's' | 'm' | 'h' | 'd';
+
+export enum LEGACY_COMPARATORS {
+  OUTSIDE_RANGE = 'outside',
+}
