+++ src/platform/packages/shared/response-ops/rule_params/metric_threshold/index.ts
@@ -1,0 +1,11 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the "Elastic License
+ * 2.0", the "GNU Affero General Public License v3.0 only", and the "Server Side
+ * Public License v 1"; you may not use this file except in compliance with, at
+ * your election, the "Elastic License 2.0", the "GNU Affero General Public
+ * License v3.0 only", or the "Server Side Public License, v 1".
+ */
+
+export { metricThresholdRuleParamsSchema } from './latest';
+export { metricThresholdRuleParamsSchema as metricThresholdRuleParamsSchemaV1 } from './v1';
