+++ x-pack/solutions/observability/plugins/infra/server/lib/alerting/common/utils.ts
@@ -5,9 +5,8 @@
  * 2.0.
  */
 
-import { isEmpty, isError } from 'lodash';
-import { schema } from '@kbn/config-schema';
-import { Logger, LogMeta } from '@kbn/logging';
+import { isError } from 'lodash';
+import type { Logger, LogMeta } from '@kbn/logging';
 import type { ElasticsearchClient } from '@kbn/core/server';
 import { ObservabilityConfig } from '@kbn/observability-plugin/server';
 import { ALERT_RULE_PARAMETERS, TIMESTAMP } from '@kbn/rule-data-utils';
@@ -53,30 +52,6 @@ const SUPPORTED_ES_FIELD_TYPES = [
   ES_FIELD_TYPES.IP,
   ES_FIELD_TYPES.BOOLEAN,
 ];
-
-export const oneOfLiterals = (arrayOfLiterals: Readonly<string[]>) =>
-  schema.string({
-    validate: (value) =>
-      arrayOfLiterals.includes(value) ? undefined : `must be one of ${arrayOfLiterals.join(' | ')}`,
-  });
-
-export const validateIsStringElasticsearchJSONFilter = (value: string) => {
-  if (value === '') {
-    // Allow clearing the filter.
-    return;
-  }
-
-  const errorMessage = 'filterQuery must be a valid Elasticsearch filter expressed in JSON';
-  try {
-    const parsedValue = JSON.parse(value);
-    if (!isEmpty(parsedValue.bool)) {
-      return undefined;
-    }
-    return errorMessage;
-  } catch (e) {
-    return errorMessage;
-  }
-};
 
 export const UNGROUPED_FACTORY_KEY = '*';
