+++ x-pack/solutions/observability/plugins/infra/server/lib/alerting/inventory_metric_threshold/register_inventory_metric_threshold_rule_type.ts
@@ -5,23 +5,11 @@
  * 2.0.
  */
 
-import { schema, Type } from '@kbn/config-schema';
 import { i18n } from '@kbn/i18n';
 import { DEFAULT_APP_CATEGORIES } from '@kbn/core/server';
 import { GetViewInAppRelativeUrlFnOpts, AlertingServerSetup } from '@kbn/alerting-plugin/server';
 import { observabilityPaths } from '@kbn/observability-plugin/common';
-import { TimeUnitChar } from '@kbn/observability-plugin/common/utils/formatters/duration';
-import {
-  InventoryItemType,
-  SnapshotMetricType,
-  SnapshotMetricTypeKeys,
-} from '@kbn/metrics-data-access-plugin/common';
-import { COMPARATORS } from '@kbn/alerting-comparators';
-import { LEGACY_COMPARATORS } from '@kbn/observability-plugin/common/utils/convert_legacy_outside_comparator';
-import {
-  SnapshotCustomAggregation,
-  SNAPSHOT_CUSTOM_AGGREGATIONS,
-} from '../../../../common/http_api';
+import { metricInventoryThresholdRuleParamsSchema } from '@kbn/response-ops-rule-params/metric_inventory_threshold';
 import type { InfraConfig } from '../../../../common/plugin_config_types';
 import { METRIC_INVENTORY_THRESHOLD_ALERT_TYPE_ID } from '../../../../common/alerting/metrics';
 import { InfraBackendLibs, InfraLocators } from '../../infra_types';
@@ -43,7 +31,6 @@ import {
   valueActionVariableDescription,
   viewInAppUrlActionVariableDescription,
 } from '../common/messages';
-import { oneOfLiterals, validateIsStringElasticsearchJSONFilter } from '../common/utils';
 import {
   createInventoryMetricThresholdExecutor,
   FIRED_ACTIONS,
@@ -52,26 +39,6 @@ import {
 } from './inventory_metric_threshold_executor';
 import { MetricsRulesTypeAlertDefinition } from '../register_rule_types';
 import { O11Y_AAD_FIELDS } from '../../../../common/constants';
-
-const comparators = Object.values({ ...COMPARATORS, ...LEGACY_COMPARATORS });
-const condition = schema.object({
-  threshold: schema.arrayOf(schema.number()),
-  comparator: oneOfLiterals(comparators) as Type<COMPARATORS>,
-  timeUnit: schema.string() as Type<TimeUnitChar>,
-  timeSize: schema.number(),
-  metric: oneOfLiterals(Object.keys(SnapshotMetricTypeKeys)) as Type<SnapshotMetricType>,
-  warningThreshold: schema.maybe(schema.arrayOf(schema.number())),
-  warningComparator: schema.maybe(oneOfLiterals(comparators)) as Type<COMPARATORS | undefined>,
-  customMetric: schema.maybe(
-    schema.object({
-      type: schema.literal('custom'),
-      id: schema.string(),
-      field: schema.string(),
-      aggregation: oneOfLiterals(SNAPSHOT_CUSTOM_AGGREGATIONS) as Type<SnapshotCustomAggregation>,
-      label: schema.maybe(schema.string()),
-    })
-  ),
-});
 
 const groupActionVariableDescription = i18n.translate(
   'xpack.infra.inventory.alerting.groupActionVariableDescription',
@@ -90,31 +57,18 @@ export function registerInventoryThresholdRuleType(
     return;
   }
 
-  const paramsSchema = schema.object(
-    {
-      criteria: schema.arrayOf(condition),
-      nodeType: schema.string() as Type<InventoryItemType>,
-      filterQuery: schema.maybe(
-        schema.string({ validate: validateIsStringElasticsearchJSONFilter })
-      ),
-      sourceId: schema.string(),
-      alertOnNoData: schema.maybe(schema.boolean()),
-    },
-    { unknowns: 'allow' }
-  );
-
   alertingPlugin.registerType({
     id: METRIC_INVENTORY_THRESHOLD_ALERT_TYPE_ID,
     name: i18n.translate('xpack.infra.metrics.inventory.alertName', {
       defaultMessage: 'Inventory',
     }),
     validate: {
-      params: paramsSchema,
+      params: metricInventoryThresholdRuleParamsSchema,
     },
     schemas: {
       params: {
         type: 'config-schema',
-        schema: paramsSchema,
+        schema: metricInventoryThresholdRuleParamsSchema,
       },
     },
     defaultActionGroupId: FIRED_ACTIONS_ID,
@@ -124,6 +78,11 @@ export function registerInventoryThresholdRuleType(
     producer: 'infrastructure',
     minimumLicenseRequired: 'basic',
     isExportable: true,
+    /*
+     * The schema defined in response-ops-rule-params cannot import all types from the plugins
+     * so the executor will expect slight differences with the type InventoryMetricThresholdParams.
+     */
+    // @ts-ignore
     executor: createInventoryMetricThresholdExecutor(libs, locators),
     actionVariables: {
       context: [
