+++ x-pack/solutions/observability/plugins/infra/server/lib/alerting/metric_threshold/register_metric_threshold_rule_type.ts
@@ -6,19 +6,17 @@
  */
 
 import { DEFAULT_APP_CATEGORIES } from '@kbn/core/server';
-import { schema } from '@kbn/config-schema';
 import { i18n } from '@kbn/i18n';
 import type {
   GetViewInAppRelativeUrlFnOpts,
   AlertingServerSetup,
 } from '@kbn/alerting-plugin/server';
 import { observabilityPaths } from '@kbn/observability-plugin/common';
-import { COMPARATORS } from '@kbn/alerting-comparators';
-import { LEGACY_COMPARATORS } from '@kbn/observability-plugin/common/utils/convert_legacy_outside_comparator';
+import { metricThresholdRuleParamsSchema } from '@kbn/response-ops-rule-params/metric_threshold';
+
 import type { InfraConfig } from '../../../../common/plugin_config_types';
 import { METRIC_THRESHOLD_ALERT_TYPE_ID } from '../../../../common/alerting/metrics';
-import { METRIC_EXPLORER_AGGREGATIONS } from '../../../../common/http_api';
-import { InfraBackendLibs, InfraLocators } from '../../infra_types';
+import type { InfraBackendLibs, InfraLocators } from '../../infra_types';
 import {
   alertDetailUrlActionVariableDescription,
   alertStateActionVariableDescription,
@@ -38,7 +36,6 @@ import {
   valueActionVariableDescription,
   viewInAppUrlActionVariableDescription,
 } from '../common/messages';
-import { oneOfLiterals, validateIsStringElasticsearchJSONFilter } from '../common/utils';
 import {
   createMetricThresholdExecutor,
   FIRED_ACTIONS,
@@ -57,57 +54,6 @@ export function registerMetricThresholdRuleType(
   if (!featureFlags.metricThresholdAlertRuleEnabled) {
     return;
   }
-  const comparator = Object.values({ ...COMPARATORS, ...LEGACY_COMPARATORS });
-  const baseCriterion = {
-    threshold: schema.arrayOf(schema.number()),
-    comparator: oneOfLiterals(comparator),
-    timeUnit: schema.string(),
-    timeSize: schema.number(),
-    warningThreshold: schema.maybe(schema.arrayOf(schema.number())),
-    warningComparator: schema.maybe(oneOfLiterals(comparator)),
-  };
-
-  const nonCountCriterion = schema.object({
-    ...baseCriterion,
-    metric: schema.string(),
-    aggType: oneOfLiterals(METRIC_EXPLORER_AGGREGATIONS),
-    customMetrics: schema.never(),
-    equation: schema.never(),
-    label: schema.never(),
-  });
-
-  const countCriterion = schema.object({
-    ...baseCriterion,
-    aggType: schema.literal('count'),
-    metric: schema.never(),
-    customMetrics: schema.never(),
-    equation: schema.never(),
-    label: schema.never(),
-  });
-
-  const customCriterion = schema.object({
-    ...baseCriterion,
-    aggType: schema.literal('custom'),
-    metric: schema.never(),
-    customMetrics: schema.arrayOf(
-      schema.oneOf([
-        schema.object({
-          name: schema.string(),
-          aggType: oneOfLiterals(['avg', 'sum', 'max', 'min', 'cardinality']),
-          field: schema.string(),
-          filter: schema.never(),
-        }),
-        schema.object({
-          name: schema.string(),
-          aggType: schema.literal('count'),
-          filter: schema.maybe(schema.string()),
-          field: schema.never(),
-        }),
-      ])
-    ),
-    equation: schema.maybe(schema.string()),
-    label: schema.maybe(schema.string()),
-  });
 
   const groupActionVariableDescription = i18n.translate(
     'xpack.infra.metrics.alerting.groupActionVariableDescription',
@@ -124,23 +70,7 @@ export function registerMetricThresholdRuleType(
     }),
     fieldsForAAD: O11Y_AAD_FIELDS,
     validate: {
-      params: schema.object(
-        {
-          criteria: schema.arrayOf(
-            schema.oneOf([countCriterion, nonCountCriterion, customCriterion])
-          ),
-          groupBy: schema.maybe(schema.oneOf([schema.string(), schema.arrayOf(schema.string())])),
-          filterQuery: schema.maybe(
-            schema.string({
-              validate: validateIsStringElasticsearchJSONFilter,
-            })
-          ),
-          sourceId: schema.string(),
-          alertOnNoData: schema.maybe(schema.boolean()),
-          alertOnGroupDisappear: schema.maybe(schema.boolean()),
-        },
-        { unknowns: 'allow' }
-      ),
+      params: metricThresholdRuleParamsSchema,
     },
     defaultActionGroupId: FIRED_ACTIONS.id,
     actionGroups: [FIRED_ACTIONS, WARNING_ACTIONS, NO_DATA_ACTIONS],
